requirements.txt
Flask
qrcode[pil]
requests
python-dotenv
